requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
alabaster==0.7.13
    # via
    #   -r requirements/base.txt
    #   sphinx
asgiref==3.7.2
    # via
    #   -r requirements/base.txt
    #   django
attrs==23.1.0
    # via
    #   -r requirements/base.txt
    #   openedx-events
babel==2.12.1
    # via
    #   -r requirements/base.txt
    #   sphinx
certifi==2023.5.7
    # via
    #   -r requirements/base.txt
    #   requests
charset-normalizer==3.1.0
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.3
    # via code-annotations
code-annotations==1.3.0
    # via -r requirements/test.in
coverage[toml]==7.2.7
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-model-utils
    #   openedx-events
django-model-utils==4.3.1
    # via -r requirements/base.txt
docutils==0.19
    # via
    #   -r requirements/base.txt
    #   sphinx
edx-opaque-keys[django]==2.3.0
    # via
    #   -r requirements/base.txt
    #   openedx-events
edx-sphinx-theme==3.1.0
    # via -r requirements/base.txt
exceptiongroup==1.1.1
    # via pytest
fastavro==1.7.4
    # via
    #   -r requirements/base.txt
    #   openedx-events
idna==3.4
    # via
    #   -r requirements/base.txt
    #   requests
imagesize==1.4.1
    # via
    #   -r requirements/base.txt
    #   sphinx
importlib-metadata==6.6.0
    # via
    #   -r requirements/base.txt
    #   sphinx
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   sphinx
markupsafe==2.1.2
    # via
    #   -r requirements/base.txt
    #   jinja2
openedx-events==8.0.1
    # via -r requirements/base.txt
packaging==23.1
    # via
    #   -r requirements/base.txt
    #   pytest
    #   sphinx
pbr==5.11.1
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.0.0
    # via pytest
pygments==2.15.1
    # via
    #   -r requirements/base.txt
    #   sphinx
pymongo==3.13.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pytest==7.3.1
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==4.1.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-slugify==8.0.1
    # via code-annotations
pytz==2023.3
    # via
    #   -r requirements/base.txt
    #   babel
    #   django
pyyaml==6.0
    # via code-annotations
requests==2.31.0
    # via
    #   -r requirements/base.txt
    #   sphinx
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   edx-sphinx-theme
snowballstemmer==2.2.0
    # via
    #   -r requirements/base.txt
    #   sphinx
sphinx==5.3.0
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   edx-sphinx-theme
sphinxcontrib-applehelp==1.0.4
    # via
    #   -r requirements/base.txt
    #   sphinx
sphinxcontrib-devhelp==1.0.2
    # via
    #   -r requirements/base.txt
    #   sphinx
sphinxcontrib-htmlhelp==2.0.1
    # via
    #   -r requirements/base.txt
    #   sphinx
sphinxcontrib-jsmath==1.0.1
    # via
    #   -r requirements/base.txt
    #   sphinx
sphinxcontrib-qthelp==1.0.3
    # via
    #   -r requirements/base.txt
    #   sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via
    #   -r requirements/base.txt
    #   sphinx
sqlparse==0.4.4
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.1.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
typing-extensions==4.6.3
    # via
    #   -r requirements/base.txt
    #   asgiref
urllib3==2.0.2
    # via
    #   -r requirements/base.txt
    #   requests
zipp==3.15.0
    # via
    #   -r requirements/base.txt
    #   importlib-metadata
